% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IonsSelect.R
\name{PlotResults}
\alias{PlotResults}
\title{PlotResults}
\usage{
PlotResults(testResults, clusterIndex)
}
\arguments{
\item{testResults}{An rMSIprocPeakMatrix object.}

\item{testResults}{Numeric Vector. A vector containing the cluster index for each pixel coded in numbers from 1 to number of clusters.}
}
\value{
Plots the results in a bar chart.
}
\description{
Plots the information related to the comparation between given clusters
}
