% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IonsSelect.R
\name{TestIonSelect}
\alias{TestIonSelect}
\title{TestIonSelect}
\usage{
TestIonSelect(PeakMtx, clusters, clusterSubset = unique(clusters),
  probability = 0.05)
}
\arguments{
\item{PeakMtx}{An rMSIprocPeakMatrix object.}

\item{clusters}{Numeric Vector. A vector containing the cluster index for each pixel coded in numbers from 1 to number of clusters.}

\item{clusterSubset}{Numeric vector containing the index of the clusters which will be avaluated.}

\item{probability}{Double. Probability restriction of the test. More information on the paper.}
}
\value{
List containing three elements. The results from the Volcano test, the results from the Zero test and the values of p, FC & Zero scores.
}
\description{
Performs a test that calculates and select the up-regulated and/or down-regulated
ions from a peak matrix between defernt clusters.
}
