% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IonsSelect.R
\name{TestIonsSelect}
\alias{TestIonsSelect}
\title{TestIonsSelect.}
\usage{
TestIonsSelect(PeakMtx, Probability, SegmentedMtx, SgmtsChoose)
}
\arguments{
\item{PeakMtx}{A matrix of peaks containg mass spectra data}

\item{Probability}{Probability restriction of the test.}

\item{SgmtsChoose}{A vector containing which clusters should be analaized.}

\item{Segmentation_matrix}{A matrix containing a segmentated peak matrix.}
}
\value{
List containing 3 elements. The results from the Volcano test, the results from the Zero test and the values of p, FC & Zero scores.
}
\description{
Performs a test that calculates and select the up-regulated and/or down-regulated
ions from a peak matrix between defernt clusters.
}
