% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IonsSelect.R
\name{TestIonSelect}
\alias{TestIonSelect}
\title{TestIonSelect}
\usage{
TestIonSelect(PeakMtx, Probability, SegmentedMtx, SgmtsChoose)
}
\arguments{
\item{PeakMtx}{A matrix of peaks containing mass spectra data}

\item{Probability}{Probability restriction of the test.}

\item{SegmentedMtx}{A matrix containing a segmentated peak matrix.}

\item{SgmtsChoose}{A vector containing which clusters should be analyzed.}
}
\value{
List containing three elements. The results from the Volcano test, the results from the Zero test and the values of p, FC & Zero scores.
}
\description{
Performs a test that calculates and select the up-regulated and/or down-regulated
ions from a peak matrix between defernt clusters.
}
